time: Using monotonic clock for local delay computation
There is lots of call in ceph where we have some code like :

  start = ceph_clock_now()
  ... do stuff ...
  end  = ceph_clock_now()
  delay = end - start
  ... do something based on delay ...

ceph_clock_now() is using the clock_gettime(CLOCK_REALTIME, ...
This calls is using the system time and could be impacted by local time
adjustement like if NTP is changing the system time.

In such case, that could lead to bad time computation like a negative count or a
huge amount of time (future time).

The willing of such code is to compute a local difference of time and not a difference of system time.

By using a monotonic clock, we only compute a difference of time. This monotonic
clocks cannot got backward. NTP adjustements have no impact on such computation.

This patch is using ceph_time, implementation initiated in commit
84dcfbd8ba6e4b630dc327c68735403c19517b69, to use monotonic clocks when its
possible. This patch adds some helpers to convert this objects into useful types
while keeping the code short & easy to read.

As a result, the local computation of time difference is much more accurate and
really represent a difference of local time.